wsgi/requirements.txt
bidict==0.23.1
Flask-Cors==4.0.0
Flask-SocketIO==5.3.6
Flask-SQLAlchemy==3.1.1
greenlet==3.0.3
h11==0.14.0
python-dotenv==1.0.1
python-engineio==4.9.0
python-socketio==5.11.1
setuptools==68.2.2
simple-websocket==1.0.0
SQLAlchemy==2.0.28
typing_extensions==4.10.0
wheel==0.41.2
wsproto==1.2.0
